requirements.txt
pytesseract
Pillow
streamlit
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
langchain-pinecone
python-dotenv
pypdf
faiss-cpu
sentence-transformers